ci: refactor release GitHub action workflow
Modified the GitHub action for generating and publishing releases. Changes enable the release workflow to be easier to maintain and more descriptive to facilitate debugging. Now the workflow first drafts a release, using the tagged 'v*' as the trigger, from there, it generates a changelog and prepares the release notes. This is done by reading the contents of the changelog into a variable and saving it back into the GitHub environment output. Afterward, a new release is created using the previously generated information.